[modify]restructure files
