+++ README.md
@@ -1,40 +1,17 @@
 # UnitTestByJest
+
 Learning Unit Test by Jest offical web site
 
-2020/3/5 DeepJavaScript 決定要讀 unit test，以 [Jest 的 Document](https://jestjs.io/docs/en/getting-started) 入手
+[Jest Docs](https://jestjs.io/docs/en/getting-started)
 
+## List
 
-## 進行方式
-
-每一個人用自己的英文名字開一個自己的分支
-
-用資料夾區分每一週的練習內容，並且寫上 README.md 說明
-
-
-每一週由導讀人先說自己做了什麼
-
-每一個人再以「與導讀人做的不同」為最優先提出，各自自由補充討論
-
-
-## 聚會
-
-- 時間: 每週四，晚上 18:30
-- 地點: 好想工作室
-
-## 目標
-
-1. 測試 JavaScript
-1. 測試切版畫面
-1. 測試前端框架
-
-## 導讀表
-
-|Date|Subject|Host|
-|--- |-------|----|
-|3/12|Getting Started|Kira|
-|3/19|Using Matchers|Lai|
-|3/26|Testing Asynchronous Code|Hello-Day|
-|4/10|Setup and Teardown|Ruru|
-|4/17|Mock Functions|Letter|
-|4/24|Jest Platform|Chris|
-|4/31|
+| Date | Subject                                                                                                |
+| ---- | ------------------------------------------------------------------------------------------------------ |
+| 3/12 | [Getting Started](https://github.com/DeepJavaScript/UnitTestByJest/tree/Kira/week1)                    |
+| 3/19 | [Using Matchers](https://github.com/DeepJavaScript/UnitTestByJest/tree/Kira/week2)                     |
+| 3/26 | [Testing Asynchronous Code](https://github.com/DeepJavaScript/UnitTestByJest/tree/helloday/2020-03-26) |
+| 4/10 | Setup and Teardown                                                                                     |
+| 4/17 | Mock Functions                                                                                         |
+| 4/24 | Jest Platform                                                                                          |
+| 4/31 |
