+++ week01_Getting_Started/README.md
@@ -55,11 +55,11 @@ module.exports = {
       '@babel/preset-env',
       {
         targets: {
-          node: 'current'
-        }
-      }
-    ]
-  ]
+          node: 'current',
+        },
+      },
+    ],
+  ],
 };
 ```
